+++ test/e2e-runner/src/linux/scenarios.rs
@@ -1,3 +1,4 @@
+pub mod all_ohis_no_service;
 pub mod ebpf_agent;
 pub mod fleet_control;
 pub mod infra_agent;